+++ packages/common/CHANGELOG.md
@@ -4,9 +4,14 @@
 
 - Cost calculation for `deployModule` transaction.
 
+### Changes
+- Function `uleb128Decode` now correctly throws an error if the given input contains more than a single ULEB128 encoded number
+
 ### Added
 
 - A `parseWallet` function to parse wallet export files
+- Function `uleb128DecodeWithIndex` that can also parse more than a single ULEB128 bigint
+- Added `parseTokenAddress` to CIS2
 
 ## 7.0.0 2023-05-15